pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "FastAPI server with UUID and cat image endpoints"
authors = ["Vaibhav-crux <vaibhavtiwari.945@gmail.com>"]
readme = "README.md"
package-mode = false  # Add this line

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"